Move compartment hook to shared CodeMirror module
